requirements/common.txt
accelerate>=1.2.1
bitsandbytes>=0.45.0
datasets>=3.2.0
gdown>=5.2.0
numpy>=1.26.0
packaging>=24.2
peft>=0.14.0
pydantic>=2.10.0
pyyaml>=6.0.2
sqlglot>=26.0.0
torch>=2.5.1
transformers>=4.49.0
trl>=0.13.0
unsloth @ git+https://github.com/unslothai/unsloth.git
unsloth_zoo @ git+https://github.com/unslothai/unsloth-zoo.git
